Abstract the detectors
